+++ lib/vmemo/application.ex
@@ -7,6 +7,10 @@ defmodule Vmemo.Application do
 
   @impl true
   def start(_type, _args) do
+    :logger.add_handler(:sentry_handler, Sentry.LoggerHandler, %{
+      config: %{metadata: [:file, :line]}
+    })
+
     children = [
       VmemoWeb.Telemetry,
       Vmemo.AshRepo,