+++ lib/vmemo_web/endpoint.ex
@@ -1,4 +1,5 @@
 defmodule VmemoWeb.Endpoint do
+  use Sentry.PlugCapture
   use Phoenix.Endpoint, otp_app: :vmemo
 
   # The session will be stored in the cookie and signed,
@@ -53,5 +54,6 @@ defmodule VmemoWeb.Endpoint do
   plug Plug.MethodOverride
   plug Plug.Head
   plug Plug.Session, @session_options
+  plug Sentry.PlugContext
   plug VmemoWeb.Router
 end